Added log observer
