+++ lib/src/core/utils/error_tracking_manager/sentry_tracking_manager.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/foundation.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sizzle_starter/src/core/utils/error_tracking_manager/error_tracking_manager.dart';
-import 'package:sizzle_starter/src/core/utils/logger.dart';
+import 'package:sizzle_starter/src/core/utils/logger/logger.dart';
+import 'package:sizzle_starter/src/feature/initialization/model/environment.dart';
 
 /// {@template sentry_tracking_manager}
 /// A class that is responsible for managing Sentry error tracking.
 /// {@endtemplate}
 final class SentryTrackingManager extends ErrorTrackingManagerBase {
   /// {@macro sentry_tracking_manager}
-  SentryTrackingManager(
-    super._logger, {
+  SentryTrackingManager({
     required this.sentryDsn,
     required this.environment,
   });
@@ -18,27 +18,21 @@ final class SentryTrackingManager extends ErrorTrackingManagerBase {
   final String sentryDsn;
 
   /// The Sentry environment.
-  final String environment;
+  final Environment environment;
 
   @override
   Future<void> report(LogMessage log) async {
     final error = log.error;
     final stackTrace = log.stackTrace;
-    final hint = log.context != null ? Hint.withMap(log.context!) : null;
 
     if (error == null && stackTrace == null) {
-      await Sentry.captureMessage(
-        log.message,
-        level: _logLevel(log.level),
-        hint: hint,
-      );
+      await Sentry.captureMessage(log.message, level: _logLevel(log.level));
       return;
     }
 
     await Sentry.captureException(
       error ?? log.message,
       stackTrace: stackTrace,
-      hint: hint,
     );
   }
 
@@ -50,7 +44,7 @@ final class SentryTrackingManager extends ErrorTrackingManagerBase {
       // Set the sample rate to 10% of events.
       options.tracesSampleRate = 0.10;
       options.debug = kDebugMode;
-      options.environment = environment;
+      options.environment = environment.value;
       options.anrEnabled = true;
       options.sendDefaultPii = true;
     });
