+++ test/src/core/utils/firebase_analytics_reporter_test.dart
@@ -4,7 +4,7 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sizzle_starter/src/core/utils/analytics/analytics_reporter.dart';
 import 'package:sizzle_starter/src/core/utils/analytics/firebase_analytics_reporter.dart';
-import 'package:sizzle_starter/src/core/utils/logger.dart';
+import 'package:sizzle_starter/src/core/utils/logger/logger.dart';
 
 @GenerateNiceMocks([MockSpec<FirebaseAnalytics>()])
 import 'firebase_analytics_reporter_test.mocks.dart';
